fix(mutasi-toko&gudang): mutation qty parsing issue

+++ resources/views/apps/mutasi/gudang-toko/index.blade.php
@@ -221,7 +221,7 @@
         $.post(link,{param:param, _token: '{{csrf_token()}}'})
             .done(function(data){
                 $("#stokbarang").html("Stok Saat Ini : "+data);
-                currentStock = parseInt(data);
+                currentStock = parseFloat(data);
             });
     }
